+++ ArchivDAT/Lösungsskizze BSRN Werkstück A.docx
